色関連の処理を共通化 Fixes #13
